% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_by_user.R
\name{gpr_fit}
\alias{gpr_fit}
\title{Gaussian process regression}
\usage{
gpr_fit(testmx, gprmodel, ncpu = -1)
}
\arguments{
\item{testmx}{Matrix; the features of testing dateset}

\item{gprmodel}{List; the output of traintraintrain(), containing four objects:
alpha
kparam
train_x
pred_method}

\item{ncpu}{Integer; the number of thread to be used;
set to -1 to use all threads; default value is -1.}
}
\value{
the prediction of testing dataset.
}
\description{
Gaussian process regression
}
