% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr_tuning.R
\name{gpr_tune}
\alias{gpr_tune}
\title{Gaussian process regression kernel parameter tuning}
\usage{
gpr_tune(x, y, kernelname = "rbf", ARD = TRUE, init_betainv = NULL,
  init_theta = NULL, optim_rbf_max = 100, optim_trace = 1,
  optim_report = 5, optim_ard_max = 50, optim_ard_trace = 1,
  optim_ard_report = 5, in_ncpu = -1)
}
\arguments{
\item{x}{Matrix; the features of tuning data set.}

\item{y}{Matrix; y.}

\item{ARD}{Boolean; set to TRUE to use ARD when tuning param;
default value is TRUE.}

\item{init_betainv}{Numeric; initial value for kernel parameter, betainv.}

\item{init_theta}{Numeric vector; initail value for kernel parameter, theta.}

\item{optim_rbf_max}{Integer; max iteration when optimizing marginal log-likelihood.}

\item{optim_trace}{Integer;}

\item{optim_report}{Integer;}

\item{optim_ard_max}{Integer;}

\item{optim_ard_trace}{Integer;}

\item{optim_ard_report}{Integer;}

\item{kernel_name}{String; kernel name}

\item{ncpu}{Integer; the number of thread to be used;
set to -1 to use all threads; default value is -1.}
}
\value{
return a list having five objects:
  betainv
  thetarel
  nfeature
  kernelname
  ARD
}
\description{
Tuning a gaussian process regression model's kernel parameter
}
