% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_by_user.R
\name{traintraintrain}
\alias{traintraintrain}
\title{Gaussian process regression}
\usage{
traintraintrain(train_x, train_y, pred_method = "sr", tune_param = TRUE,
  ARD = TRUE, kname = "gaussiandotrel", ktheta = NULL, kbetainv = NULL,
  ncpu = -1, srsize = NULL, tsize = NULL, clus_size = NULL)
}
\arguments{
\item{train_x}{Matrix; the features of training data set.}

\item{train_y}{Matrix; y.}

\item{pred_method}{String; Set the model training approach.
cg_direct_lm: conjugate gradient decent with no preconditioning
usebigK: use built-in matrix inversion
local_gpr: use Local GPR method
sr: subset regressors}

\item{tune_param}{Boolean; Set to TRUE to tune parameters of kernel function
default value is TRUE.}

\item{ARD}{Boolean; set to TRUE to use ARD when tuning param;
default value is TRUE.}

\item{kname}{String; the name of kernel; default value is 'gaussiandotrel'.}

\item{ktheta}{Numeric vector; store kernel parameter;
should be provided when tune_param is FALSE.}

\item{kbetainv}{Numeric; store kernel parameter betainv;
shuld be provided when tune_param is FALSE.}

\item{ncpu}{Integer; the number of thread to be used;
set to -1 to use all threads; default value is -1.}

\item{srsize}{Integer; the size of subtraining dataset;
should be provided when pred_method = 'sr'}

\item{tsize}{Integer; parameter used for model tuning (tune_param = TRUE),
only tsize of training points will be used for model tuning}

\item{clus_size}{Integer; parameter for local_gpr: set the cluster size;
should be provided when pred_method = 'local_gpr'.}
}
\value{
return a list having four objects:
  alpha
  kparam
  train_x
  pred_method
}
\description{
Training a gaussian process regression model
}
