% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_gpr.R
\name{gbm_train}
\alias{gbm_train}
\title{Gradient boosting machine based on GPR}
\usage{
gbm_train(train_x, train_y, test_x, test_y, pred_method = "2",
  n_model = 500, batch_size = 1000, lr = 0.1, decay_lr = 1,
  tune_param = FALSE, tune_size = NULL, sr_size = NULL,
  update_kparam_tiems = 50, update_col_sample = 50, update_lr = 50,
  kname = "gaussiandotrel", ktheta = NULL, kbetainv = NULL, ncpu = -1)
}
\arguments{
\item{train_x}{Matrix; the features of training data set.}

\item{train_y}{Matrix; y.}

\item{pred_method}{String; Set the model training approach.
1: random row sampling after all training data have been used.
2: random row sampling.
3: row sampling plus col sampling.}

\item{n_model}{Positive number of submodel in gbm.}

\item{batch_size}{Positive integer; batch size for each iteration.}

\item{lr}{Numeric between 0-1; learning rate.}

\item{decay_lr}{Numeric between 0-1; decay learning rate, default = 1 (no decay).}

\item{tune_param}{Boolean; Set to TRUE to tune parameters of kernel function
default value is TRUE.}

\item{tune_size}{Positve integer; size of tuning data set.}

\item{sr_size}{Positive integer; size of sub dataset in gbm_sr}

\item{update_col_sample}{Positive integer; time to update kernel parameter(for method 3).}

\item{update_lr}{Positive integer; time to decay learning rate; default is 50.}

\item{kname}{String; the name of kernel; default value is 'gaussiandotrel'.}

\item{ktheta}{Numeric vector; store kernel parameter;
should be provided when tune_param is FALSE.}

\item{kbetainv}{Numeric; store kernel parameter betainv;
shuld be provided when tune_param is FALSE.}

\item{ncpu}{Integer; the number of thread to be used;
set to -1 to use all threads; default value is -1.}

\item{tune_param}{Boolean; Set to TRUE to tune parameters of kernel function,
default value for pred_method 1 & 2 is TRUE.
default value for pred_method 3 is FALSE.}

\item{update_kparam_times}{Positve integer; time to update kernel parameter(for method 1/2).}
}
\value{
return a list having four objects:
  models
  pred_method
  train_rmse
  test_rmse
}
\description{
Training a gradient boosting machine
}
